+++ Park/src/main/kotlin/models/TicketBooth.kt
@@ -6,7 +6,7 @@ import java.time.LocalDateTime
 
 class TicketBooth {
     private var ticketCounter: Long = 0
-    fun getTicket(floorNumber: Long, spotNumber: Long, entryTime: LocalDateTime): Ticket {
+    fun getTicket(floorNumber: Int, spotNumber: Int, entryTime: LocalDateTime): Ticket {
         ticketCounter++
 
         return NormalTicket(
