+++ Park/src/main/kotlin/models/receipts/NormalReceipt.kt
@@ -4,8 +4,8 @@ import java.time.LocalDateTime
 
 data class NormalReceipt(
     val receiptNumber: Long,
-    val floorNumber: Long,
-    val spotNumber: Long,
+    val floorNumber: Int,
+    val spotNumber: Int,
     val entryDateTime: LocalDateTime,
     val fee: Long,
     val exitDateTime: LocalDateTime
