+++ Park/src/main/kotlin/models/tickets/NormalTicket.kt
@@ -4,19 +4,19 @@ import java.time.LocalDateTime
 
 data class NormalTicket(
     private val ticketNumber: Long,
-    private val floorNumber: Long,
-    private val spotNumber: Long,
+    private val floorNumber: Int,
+    private val spotNumber: Int,
     private val entryDateTime: LocalDateTime
 ) : Ticket {
     override fun getTicketEntryDateTime(): LocalDateTime {
         return entryDateTime
     }
 
-    override fun getSpotNumberForTicket(): Long {
+    override fun getSpotNumberForTicket(): Int {
         return spotNumber
     }
 
-    override fun getFloorNumberForTicket(): Long {
+    override fun getFloorNumberForTicket(): Int {
         return floorNumber
     }
